Extract export metadata helpers
