+++ apps/frontend/src/lib/exportTypes.ts
@@ -1,0 +1,3 @@
+export type ExportFormat = "mp4" | "webm" | "mov" | "mkv";
+
+export type ExportResolution = "original" | "1080" | "720";
